refactor(grow): extract early LLM phases to llm_phases.py
Move phases 2, 3, 4a-4f from stage.py to llm_phases.py as methods
on _LLMPhaseMixin. GrowStage inherits from both _LLMHelperMixin
and _LLMPhaseMixin, preserving all self._* references.

llm_phases.py: 1,142 lines (8 phase methods)
stage.py: 1,319 lines (down from 2,420)

Phases 8c, 9, 9b, 9c remain in stage.py for the next PR.

Part of #851.
